Implemented JwtFilter which intercepts every incoming HTTP requests
It checks if the incoming request contains an Authorization header with a value starting with "Bearer ", if not it passes the request along the filter chain without further processing otherwise it extracts the username and all the claims from the token.
It then loads the user details if the user is not yet authenticated.
If the token is valid, it creates an Authentication object of type UsernamePasswordAuthenticationToken, sets it in the SecurityContextHolder, and finally it continues the filter chain.
The filter also provides methods such as isAdmin(), isUser() and getCurrentUser() to check the role and get the current user's email based on the claims extracted from the JWT token.