Created service CustomUserDetailsService that implements UserDetailsService
The loadUserByUsername method is implemented to load user details based on the provided username from the UserDao by calling findByEmail(email).
If a user with the given email is found, it constructs a UserDetails object which represents the authenticated user (containing the username, password, and authorities) or it throws a UsernameNotFoundException.