Created the dashboard service that implements the HTTP call to the /dashboard/details endpoint
In previous calls to the backend, authentication was not required for the endpoints. Now, I've added the jwt token to the http headers by invoking the retrieveToken method in the auth service. This allows the backend's JwtFilter class, which extends OncePerRequestFilter, to verify the validity and expiration of the token before proceeding to resolve the request.