Differentiated user and admin dashboards
Renamed the old dashboard-page component to dashboard-page-admin. Implemented an if condition in the login-page.component.ts file to check the role in the token by calling the retrieveTokenRole() method in AuthService and redirect to the appropriate dashboard page.
In the app-routing module, set the expectedRole to 'admin' for the category and item pages.