Update Bill and Item classes and services
Added the issueDate field to the Bill class for database storage and inclusion in the bill pdf.
Changed the variable type of price and total from Integer to Double in the ItemDto, Item, and Bill classes.
Updated prepareData in the BillServiceImpl class and added setDateForm method.
Made necessary adjustments in the saveBill method (in BillServiceImpl) and in the getItemFromMap method (in ItemServiceImpl) to accommodate the changes in variable types.