In users-page component, implemented the functionality to toggle user status via slide-toggle element
In the html template, I added to the table listing users their current status displayed and a slide-toggle allowing status activation or deactivation.
In the UsersPage component, the onStatusChange method handles the status change event by updating the status through a service call to updateStatus.
The updateStatus method in the UserService sends an http post request to the server endpoint /users/updateStatus with the required data (user's status and ID).